+++ Magic/CHANGELOG.md
@@ -9,6 +9,7 @@
  - Add "angry" option for wolf mobs, default interval to 1 second if not specified and using interval triggers
  - Add Equation action .. not sure what you'd use this for, but there is a spiral example in elworld/spells
  - Added Sound and Particle effectlib effects, and "probability" parameter
+ - The Tame action/spell can now be used to make foxes trust you
 
 # 7.10.3